refactor: Config 클래스 support/config 이동
